hf_requirements.txt
torch==2.2.2
torchvision==0.17.2
timm==0.9.16
gradio==4.44.0
Pillow==10.4.0
numpy==1.26.4
